data prep optimized

+++ prepare_global_features.py
@@ -2,6 +2,7 @@ import os
 import numpy as np
 from PIL import Image
 from annotator.content import ContentDetector
+from tqdm import tqdm
 
 # Initialize the ContentDetector
 apply_content = ContentDetector()
@@ -29,10 +30,15 @@ def save_embedding(image_path, output_dir):
     - image_path (str): Path to the image file.
     - output_dir (str): Directory to save the .npy file.
     """
-    content_emb = process_image(image_path)
     image_name = os.path.basename(image_path)
     npy_filename = os.path.splitext(image_name)[0] + '.npy'
     npy_path = os.path.join(output_dir, npy_filename)
+    
+    if os.path.exists(npy_path):
+        # print(f"File already exists, skipping: {npy_path}")
+        return
+
+    content_emb = process_image(image_path)
     np.save(npy_path, content_emb)
     print(f"Saved: {npy_path}")
 
@@ -47,7 +53,7 @@ def process_images_in_folder(input_folder, output_folder):
     if not os.path.exists(output_folder):
         os.makedirs(output_folder)
 
-    for filename in os.listdir(input_folder):
+    for filename in tqdm(os.listdir(input_folder)):
         if filename.endswith(('.png', '.jpg', '.jpeg')):
             image_path = os.path.join(input_folder, filename)
             save_embedding(image_path, output_folder)
